+++ changelog.md
@@ -5,6 +5,7 @@ TGUI 1.3  (TBD)
 - New widget: EditBoxSlider ([PR #238](https://github.com/texus/TGUI/pull/238))
 - All widgets can now be configured to ignore mouse events
 - Added HorizontalLayout and VerticalLayout to replace widget-specific enums
+- Added method to associate user data to combo box items
 - Renamed isKeptInParent to getKeepInParent in ChildWindow
 - Renamed limitTextWidth to setTextWidthLimited in EditBox
 - String::fromNumber now supports int8_t (but no longer accepts pointers)